task 12 - German cried while selling (do poprawki)

+++ src/com/company/devices/Car.java
@@ -2,13 +2,22 @@ package com.company.devices;
 
 import com.company.creatures.Human;
 
+import java.sql.Array;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.ExecutionException;
 
 public abstract class Car extends Device implements Comparable<Car> {
     public Double value;
+    public ArrayList<Human> owners;
+    public ArrayList<Integer> transactions;
+    public Integer transact;
+    static Integer DEFAULT_NUMBER_OF_TRANSACTIONS = 0;
+
 
     public Car(String model, String producent, Integer yearOfProduction, Double price) {
         super();
+        this.transact = DEFAULT_NUMBER_OF_TRANSACTIONS+1;
     }
 
     @Override
@@ -36,6 +45,10 @@ public abstract class Car extends Device implements Comparable<Car> {
         if (buyer.getCash() < price){
             throw new Exception("Not enough money.");
         }
+        if (!seller.haveCar(this) && !seller.owner(this)){
+            throw new Exception("This car is stolen");
+        }
+
 
         seller.deleteCar(this);
         buyer.addCar(this);
@@ -49,5 +62,39 @@ public abstract class Car extends Device implements Comparable<Car> {
     @Override
     public int compareTo(Car o){
         return (this.yearOfProduction-o.yearOfProduction);
+    }
+
+    public ArrayList<Human> getOwners(){
+        return owners;
+    }
+
+    public Human getCurrentOwner(){
+        return this.owners.get(this.owners.size()-1);
+    }
+
+    public void checkOwner (Human owner){
+        if (this.owners.contains(owner)){
+            for (Human owners : getOwners()) {
+                System.out.println(owners.firstName + "" + owners.lastName + " ");
+            }
+        }else {
+            System.out.println("No owner.");
+        }
+    }
+
+    public boolean wasOwner(Human owner){
+        return this.owners.contains(owner);
+    }
+
+    public void humanAsoldToHumanB(Human a, Human b) {
+        if (wasOwner(a) && wasOwner(b)){
+            System.out.println("Was sold.");
+        }else {
+            System.out.println("Wasn't sold.");
+        }
+    }
+
+    public int numberOfTransactionOnThisCar(){
+        return this.transactions.size();
     }
 }
